% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alocv_lasso.R
\name{alo_glmnet}
\alias{alo_glmnet}
\title{Fits and computes the approximate leave-one-out cross validation for glmnet.}
\usage{
alo_glmnet(
  x,
  y,
  family = c("gaussian", "binomial", "poisson"),
  weights = NULL,
  offset = NULL,
  alpha = 1,
  nlambda = 100,
  lambda.min.ratio = ifelse(nobs < nvars, 0.01, 1e-04),
  lambda = NULL,
  standardize = TRUE,
  intercept = TRUE,
  thresh = 1e-07,
  dfmax = nvars + 1,
  ...,
  lasso_approximate_intercept = TRUE
)
}
\arguments{
\item{x}{The data matrix.}

\item{y}{The response vector.}

\item{family}{The family of the glm to fit.}

\item{weights}{Observation weights.}

\item{offset}{A vector of length nobs included in the linear predictor.}

\item{alpha}{Elastic net parameter.}

\item{nlambda}{The number of \code{lambda} values.}

\item{lambda.min.ratio}{Smallest value for \code{lambda} as a fraction of the largest.}

\item{lambda}{A user supplied lambda sequence.}

\item{standardize}{Whether to apply \code{x} variable standardization.}

\item{intercept}{Whether intercepts should be fitted.}

\item{thresh}{Convergence threshold for coordinate descent.}

\item{dfmax}{Limit maximum number of variables in the model.}

\item{...}{Other parameters passed on to \code{\link[glmnet]{glmnet}}}

\item{lasso_approximate_intercept}{If true, uses a slightly less accurate but
faster method for the pure LASSO case (which ignores the variability of the
intercept).}
}
\value{
An object with S3 class \code{"alo"} in addition to the
classes corresponding to the fitted object returned by \code{\link[glmnet]{glmnet}}.
Additional fields are introduced to the object to represent the computed ALO
values.
\item{alo}{A numerical vector representing the ALO deviance loss for each tuning value.}
\item{alo_mse}{A numerical vector representing the ALO MSE loss for each tuning value.}
\item{alo_mae}{A numerical vector representing the ALO MAE loss for each tuning value.}
\item{leverage}{A numerical matrix representing the computed leverage of each data point for each tuning value.}
}
\description{
Fits and computes the approximate leave-one-out cross validation for glmnet.
}
\seealso{
\code{\link[glmnet]{glmnet}}
}
