% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alocv_lasso.R
\name{alo_glmnet_internal}
\alias{alo_glmnet_internal}
\title{Internal code to compute ALO for Lasso / elastic net problems.}
\usage{
alo_glmnet_internal(
  x,
  beta,
  y,
  lambda,
  family,
  alpha,
  a0,
  standardize,
  intercept,
  lasso_approximate_intercept = FALSE
)
}
\arguments{
\item{x}{The data matrix}

\item{beta}{A matrix of fitted parameter values for each tuning value.}

\item{y}{The response vector.}

\item{lambda}{A vector of penalization values.}

\item{family}{The family of the glm that was fit.}

\item{alpha}{The elastic-net parameter.}

\item{a0}{The fitted intercept values for each penalization value.}

\item{standardize}{A logical value indicating whether \code{x} should be standardized.}

\item{intercept}{A logical value indicating whether an intercept was fit to the model.}

\item{lasso_approximate_intercept}{If true, uses a faster algorithm for the pure lasso
case even when the intercept is given. This ignores the contribution of the uncertainty
of the intercept to the final estimation error.}
}
\description{
Internal code to compute ALO for Lasso / elastic net problems.
}
