% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alo_lasso_rcpp}
\alias{alo_lasso_rcpp}
\title{Computes alo values for the LASSO problem using a up/down-dating
algorithm to reuse intermediate values along a given regularization
path.}
\usage{
alo_lasso_rcpp(A, B, y, a0 = NULL)
}
\arguments{
\item{A}{The predictor matrix (\code{n x p}).}

\item{B}{The matrix of fitted values for each regularization (\code{p x m})}

\item{y}{The vector of observed responses (length \code{n}).}

\item{a0}{An optional vector of fitted intercepts.}
}
\value{
A list with the computed alo values and leverage.
}
\description{
Note that due to the special algorithm, it is not strictly correct
for models fitted with an intercept, as it does not account for
the uncertaining in estimating the intercept.
}
