% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alocv_svm.R
\name{alo_svm}
\alias{alo_svm}
\title{Fits and computes the approximate cross-validation for SVMs.}
\usage{
alo_svm(x, y, scale = TRUE, type = NULL, kernel = "radial",
  degree = 3, gamma = if (is.vector(x)) 1 else 1/ncol(x), coef0 = 0,
  cost = 1, nu = 0.5, tolerance = 1e-04, use_rfp = TRUE,
  use_pivot = FALSE, ...)
}
\arguments{
\item{x}{The data matrix}

\item{y}{The response vector}

\item{scale}{A logical vector indicating the variables to be scaled.}

\item{type}{The type of svm to fit. Only 'C-classification' is supported at the moment.}

\item{kernel}{The kernel used in traaining and predicting.}

\item{degree}{Parameter for polynomial kernels.}

\item{gamma}{Scale parameter needed for all kernels. Default 1 / (data dimension).}

\item{coef0}{Offset parameter needed for kernels.}

\item{cost}{Cost of constrain violations.}

\item{nu}{Hyperparameter for some SVM types.}

\item{tolerance}{Tolerance for SVM solver, and for support vector detection in ALO procedure.}

\item{use_rfp}{If true, uses rectangular full packed matrices in ALO algorithm. Reduces memory requirement.}

\item{use_pivot}{If true, uses pivoting in Cholesky factorization of kernel matrix.}

\item{...}{Additional parameters to be passed to the underlying SVM fit function.}
}
\description{
Fits and computes the approximate cross-validation for SVMs.
}
\seealso{
\code{\link[e1071]{svm}}
}
