% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alocv_svm.R
\name{alo_svm_internal}
\alias{alo_svm_internal}
\title{Approximate Leave-one-out cross validation for SVM objects}
\usage{
alo_svm_internal(x, y, alpha, rho, lambda, kernel = "radial",
  gamma = if (is.vector(x)) 1 else 1/ncol(x), degree = 3, coef0 = 0,
  tolerance = 1e-04, use_rfp = TRUE, use_pivot = FALSE)
}
\arguments{
\item{x}{The data matrix.}

\item{y}{The response vector.}

\item{tolerance}{The tolerance to use in detecting support vectors.}

\item{use_rfp}{Whether to use rectangular full packed format to represent the kernel. Reduces memory usage.}

\item{use_pivot}{Whether to use pivoted Cholesky decomposition. Set this to true when the kernel
is not strictly positive-definite.}

\item{fit}{The fitted svm object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Approximate Leave-one-out cross validation for SVM objects
}
