% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alo_lasso.R
\name{alocv.elnet}
\alias{alocv.elnet}
\title{Computes the approximate leave one out estimator for elastic-net
estimators.}
\usage{
\method{alocv}{elnet}(fit)
}
\description{
This function computes the approximate leave-one-out risk and leverage
along a path produced by a call to `glmnet`.
}
