% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alocv_lasso.R
\name{alocv.glmnet}
\alias{alocv.glmnet}
\title{Computes approximate leave-one-out errors for each regularization
value in the fitted object.}
\usage{
\method{alocv}{glmnet}(
  fit,
  x,
  y,
  alpha = NULL,
  standardize = NULL,
  intercept = NULL,
  ...,
  lasso_approximate_intercept = TRUE
)
}
\arguments{
\item{fit}{the fitted glmnet object}

\item{x}{the predictor matrix}

\item{y}{the response vector}

\item{alpha}{the elastic-net parameter used to fit the model.}

\item{standardize}{indicates whether \code{\link[glmnet]{glmnet}} was called with standardize.
If not specified, this method will try to guess by examining the fitted object.}

\item{intercept}{indicates whether \code{\link[glmnet]{glmnet}} was called with an intercept.
If not specified, this method will try to guess by examining the fitted object.}

\item{...}{Further arguments passed to or from other methods.}

\item{lasso_approximate_intercept}{If true, uses a slightly less accurate but
faster method for the pure LASSO case (which ignores the variability of the
intercept).}
}
\description{
Computes approximate leave-one-out errors for each regularization
value in the fitted object.
}
\seealso{
\code{\link[glmnet]{glmnet}}
}
