% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alocv.R
\docType{package}
\name{alocv}
\alias{alocv}
\alias{alocv-package}
\title{alocv: A package for computing unbiased risk estimates for high-dimensional models.}
\description{
alocv: A package for computing unbiased risk estimates for high-dimensional models.
}
