% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alocv_svm.R
\name{alocv.svm}
\alias{alocv.svm}
\title{Approximate Leave-one-out cross validation for SVM objects}
\usage{
\method{alocv}{svm}(fit, x, y, tolerance = 1e-05, use_rfp = TRUE, use_pivot = FALSE, ...)
}
\arguments{
\item{fit}{The fitted svm object.}

\item{x}{The data matrix.}

\item{y}{The response vector.}

\item{tolerance}{The tolerance to use in detecting support vectors.}

\item{use_rfp}{Whether to use rectangular full packed format to represent the kernel. Reduces memory usage.}

\item{use_pivot}{Whether to use pivoted Cholesky decomposition. Set this to true when the kernel
is not strictly positive-definite.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Approximate Leave-one-out cross validation for SVM objects
}
