% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alocv_lasso.R
\name{check_standardize}
\alias{check_standardize}
\title{Given a glm fit and the data, determines whether
standardization was applied to the data before
the fit.}
\usage{
check_standardize(fit, x, y, alpha)
}
\arguments{
\item{fit}{The fitted object}

\item{x}{The data matrix}

\item{y}{The response vector}

\item{alpha}{The elastic net parameter.}
}
\description{
This check is done by evaluating the dual optimal
value and checking whether it corresponds to the
stated penalization value.
}
