% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alocv_plot.R
\name{plot.alo_glmnet}
\alias{plot.alo_glmnet}
\title{Plots information for the given ALO evaluation.}
\usage{
\method{plot}{alo_glmnet}(x, ...)
}
\arguments{
\item{x}{The alo object to plot.}

\item{...}{Additional arguments to be passed to the underlying plotting function.}
}
\description{
Plots information for the given ALO evaluation.
}
