% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alocv.R
\docType{package}
\name{alocv}
\alias{alocv}
\alias{alocv-package}
\alias{alocv}
\title{alocv: A package for computing unbiased risk estimates for high-dimensional models.}
\usage{
alocv(fit)
}
\arguments{
\item{fit}{An object containing the fitted estimator.}
}
\description{
alocv: A package for computing unbiased risk estimates for high-dimensional models.

Computes an approximate-leave-one out risk for the given fit.
}
